workers/audio_generation/main.py
from ..base import BaseWorker
from sqlalchemy.orm import Session
from db.session import SessionLocal
from models import AudioGenerationJob, JobStatus, Voice
from core.config import settings
import logging
from datetime import datetime
import sys
import io
from utils.s3 import load_file_from_s3, upload_file_to_s3
import os
import shutil

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)

class AudioGenerator(BaseWorker):
    def __init__(self):
        self.base_file_path = "temp_audio_generation"
        os.makedirs(self.base_file_path, exist_ok=True)

        logger.info(f"Initializing AudioGenerator with queue: {settings.AUDIO_GENERATION_QUEUE}")
        
        # TODO: Initialize audio generation models/services here
        # Example: self.tts_model = load_tts_model()
        # Example: self.audio_processor = AudioProcessor()
        
        super().__init__(settings.AUDIO_GENERATION_QUEUE)

    def generate_audio(self, job_data: dict):
        """Generate audio for a given job"""
        job_id = job_data.get("job_id")
        text_content = job_data.get("text_content")
        voice_id = job_data.get("voice_id")
        metadata = job_data.get("metadata", {})

        logger.info(f"Generating audio - Job ID: {job_id}, Voice ID: {voice_id}")
        
        # TODO: Implement actual audio generation logic here
        # Example steps:
        # 1. Load voice model/embedding for the specified voice_id
        # 2. Process text content (text normalization, etc.)
        # 3. Generate audio using TTS model with voice embedding
        # 4. Post-process audio (format conversion, quality enhancement, etc.)
        # 5. Save generated audio to temporary file
        
        # Placeholder for generated audio file path
        temp_audio_path = os.path.join(self.base_file_path, f"generated_audio_{job_id}.wav")
        
        # TODO: Replace with actual audio generation
        # Example: self.tts_model.generate_audio(text_content, voice_embedding, temp_audio_path)
        
        logger.info(f"Audio generated and saved to {temp_audio_path}")
        
        # TODO: Upload generated audio to S3
        # Example: s3_key = f"generated_audio/{job_id}/{os.path.basename(temp_audio_path)}"
        # Example: upload_file_to_s3(temp_audio_path, s3_key)
        
        # TODO: Clean up temporary files
        # Example: os.remove(temp_audio_path)
        
        return {
            "audio_file_path": temp_audio_path,
            "s3_key": "placeholder_s3_key",  # TODO: Replace with actual S3 key
            "duration": 0.0,  # TODO: Calculate actual duration
            "format": "wav"  # TODO: Get actual format
        }

    def process(self, job_data: dict):
        """Process an audio generation job"""
        job_id = job_data.get("job_id")
        text_content = job_data.get("text_content")
        voice_id = job_data.get("voice_id")
        
        if not job_id or not text_content or not voice_id:
            logger.error(f"Invalid message format: {job_data}")
            return

        logger.info(f"Processing audio generation job {job_id} with data: {job_data}")
        db = SessionLocal()
        job = None
        
        try:
            # TODO: Get the job from database
            # Example: job = db.query(AudioGenerationJob).filter_by(id=job_id, is_deleted=False).first()
            # if not job:
            #     logger.error(f"Job {job_id} not found")
            #     return

            # TODO: Update job status to processing
            # Example: job.status = JobStatus.PROCESSING
            # Example: db.commit()
            
            # Generate the audio
            result = self.generate_audio(job_data)
            
            # TODO: Update job status to completed with results
            # Example: job.status = JobStatus.COMPLETED
            # Example: job.result = {
            #     "message": "Audio generation completed successfully",
            #     "voice_id": voice_id,
            #     "audio_file_path": result["audio_file_path"],
            #     "s3_key": result["s3_key"],
            #     "duration": result["duration"],
            #     "format": result["format"],
            #     "processed_at": datetime.utcnow().isoformat()
            # }
            # Example: db.commit()
            
            logger.info(f"Successfully processed audio generation job {job_id}")
            
        except Exception as e:
            logger.error(f"Error processing audio generation job {job_id}: {str(e)}")
            if job:
                try:
                    # TODO: Update job status to failed
                    # Example: job.status = JobStatus.FAILED
                    # Example: job.error = str(e)
                    # Example: db.commit()
                    pass
                except Exception as commit_error:
                    logger.error(f"Failed to update job status: {str(commit_error)}")
        finally:
            db.close()

if __name__ == "__main__":
    logger.info("Starting AudioGenerator worker...")
    generator = AudioGenerator()
    try:
        logger.info("Starting to consume messages...")
        generator.start()
    except KeyboardInterrupt:
        logger.info("Received keyboard interrupt, shutting down...")
        generator.close()
    except Exception as e:
        logger.error(f"Unexpected error: {str(e)}")
        generator.close()
