workers/audio_generation/requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
audioread==3.0.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
conformer==0.3.2
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
executing==2.2.0
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.18.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
resemble-perth==1.0.1
rich==14.0.0
s3tokenizer==0.1.7
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.2.0
typer==0.16.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
zipp==3.23.0
pika==1.3.2